add settings to nodemetamodel
